without mem _3

+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: akrid <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/01 16:03:27 by akrid             #+#    #+#             */
-/*   Updated: 2023/11/02 15:11:32 by akrid            ###   ########.fr       */
+/*   Updated: 2023/11/03 16:30:11 by akrid            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
